+++ src/test/java/com/siberalt/singularity/presenter/google/series/CandlePriceSeriesTest.java
@@ -92,7 +92,7 @@ class CandlePriceSeriesTest {
         );
 
         CandlePriceSeriesProvider series = new CandlePriceSeriesProvider(candles)
-            .setPriceExtractor(c -> c.getClosePrice().toBigDecimal().doubleValue() * 2);
+            .setPriceExtractor(c -> c.getClosePrice().toDouble() * 2);
 
         Optional<SeriesChunk> result = series.provide(0, 1, 1);
